xn--d1abkig.online/pyproject.toml
[tool.poetry]
name = "xn-d1abkig-online"
version = "0.1.0"
description = ""
authors = ["Krompir12 <fauxvotredeces@gmail.com>"]
readme = "README.md"
packages = [{include = "xn_d1abkig"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.9"
django = "4.1.10"
pillow = "10.0.0"
python-dotenv = "1.0.0"
pytz = "2023.3"
requests = "2.31.0"
black = "23.7.0"
sentry-sdk = "^1.34.0"
djangorestframework = "^3.14.0"
mysqlclient = "2.2.0"
django-cors-headers = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 79
py_version = 310
